requirements.txt
geopandas
shapely
pyproj
osmnx
h3
pandas
numpy
requests
scikit-learn
xgboost
folium
joblib
pyyaml
python-dotenv
tqdm
streamlit
streamlit-folium
jsonschema>=4.17.0
